# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Callora API",
    "version": "1.0.0",
    "description": "API contract covering Callora backend billing, usage, and developer routes."
  },
  "paths": {
...
    "/api/quota/requests": {
      "post": {
        "summary": "Submit a quota request",
        "description": "Creates a new quota increase request for the authenticated developer. The request remains pending until an admin resolves it.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotaRequestCreate"
              },
              "examples": {
                "createRequest": {
                  "summary": "Quota increase request",
                  "value": {
                    "requested_tier": "pro",
                    "reason": "We need a higher monthly call cap for our beta launch.",
                    "requested_overrides": {
                      "monthly_call_limit": 250000,
                      "rate_limit_max_requests": 1000
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quota request created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuotaRequest"
                    }
                  }
                },
                "examples": {
                  "createdRequest": {
                    "summary": "Created quota request",
                    "value": {
                      "data": {
                        "id": "8dbc1b8b-5fe5-4e2f-8ece-9f0d7f9e5743",
                        "developerId": "dev_123",
                        "requestedTier": "pro",
                        "reason": "We need a higher monthly call cap for our beta launch.",
                        "requestedOverrides": {
                          "monthlyCallLimit": 250000,
                          "rateLimitMaxRequests": 1000
                        },
                        "status": "pending",
                        "createdAt": "2026-07-26T12:34:56.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "validationError": {
                    "summary": "Validation error",
                    "value": {
                      "code": "VALIDATION_ERROR",
                      "message": "Request validation failed",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Missing or invalid authentication",
                    "value": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication required",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unprocessable": {
                    "summary": "Unprocessable entity",
                    "value": {
                      "code": "UNPROCESSABLE_ENTITY",
                      "message": "Request validation failed",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List quota requests",
        "description": "Lists the authenticated developer's own quota requests. An optional status query parameter filters the results.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Optional status filter",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of quota requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/QuotaRequest"
                      }
                    }
                  }
                },
                "examples": {
                  "listSuccess": {
                    "summary": "List of requests",
                    "value": {
                      "data": [
                        {
                          "id": "8dbc1b8b-5fe5-4e2f-8ece-9f0d7f9e5743",
                          "developerId": "dev_123",
                          "requestedTier": "pro",
                          "reason": "We need a higher monthly call cap for our beta launch.",
                          "requestedOverrides": {
                            "monthlyCallLimit": 250000,
                            "rateLimitMaxRequests": 1000
                          },
                          "status": "pending",
                          "createdAt": "2026-07-26T12:34:56.000Z"
                        },
                        {
                          "id": "c51082b2-9f67-4f57-8d7e-0784f09e5ea6",
                          "developerId": "dev_123",
                          "requestedTier": "enterprise",
                          "reason": "We are onboarding a larger enterprise customer base.",
                          "status": "approved",
                          "adminNotes": "Approved for the next billing cycle.",
                          "resolvedBy": "admin_42",
                          "resolvedAt": "2026-07-27T09:00:00.000Z",
                          "createdAt": "2026-07-25T10:15:00.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidStatus": {
                    "summary": "Invalid status value",
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "status must be one of: pending, approved, rejected",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Missing or invalid authentication",
                    "value": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication required",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/quota/requests/{id}": {
      "get": {
        "summary": "Get a quota request",
        "description": "Returns a single quota request by ID for the authenticated developer. Requests belonging to another developer appear as not found.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Quota request UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quota request retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuotaRequest"
                    }
                  }
                },
                "examples": {
                  "singleRequest": {
                    "summary": "Single quota request",
                    "value": {
                      "data": {
                        "id": "8dbc1b8b-5fe5-4e2f-8ece-9f0d7f9e5743",
                        "developerId": "dev_123",
                        "requestedTier": "pro",
                        "reason": "We need a higher monthly call cap for our beta launch.",
                        "requestedOverrides": {
                          "monthlyCallLimit": 250000,
                          "rateLimitMaxRequests": 1000
                        },
                        "status": "pending",
                        "createdAt": "2026-07-26T12:34:56.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Missing or invalid authentication",
                    "value": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication required",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Quota request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notFound": {
                    "summary": "Quota request missing or inaccessible",
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "Quota request not found",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-api-key"
      }
    },
    "schemas": {
...
      "QuotaRequest": {
        "type": "object",
        "description": "A developer quota increase request and its current resolution state.",
        "required": [
          "id",
          "developerId",
          "requestedTier",
          "reason",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique request ID (UUID v4)"
          },
          "developerId": {
            "type": "string",
            "description": "The developer's user ID (from auth)"
          },
          "requestedTier": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ],
            "description": "Requested plan tier"
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "description": "Developer-provided justification"
          },
          "requestedOverrides": {
            "type": "object",
            "description": "Optional specific limits to override in addition to the tier",
            "properties": {
              "monthlyCallLimit": {
                "type": "integer",
                "minimum": 1,
                "description": "Requested monthly API call cap"
              },
              "rateLimitMaxRequests": {
                "type": "integer",
                "minimum": 1,
                "description": "Requested per-window rate-limit ceiling"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "description": "Current resolution state"
          },
          "adminNotes": {
            "type": "string",
            "description": "Admin-provided notes on the approval or rejection decision"
          },
          "resolvedBy": {
            "type": "string",
            "description": "Actor ID (admin) who resolved the request"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 timestamp when the request was resolved"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 timestamp when the request was submitted"
          }
        }
      },
      "QuotaRequestCreate": {
        "type": "object",
        "description": "Payload for submitting a new quota increase request.",
        "required": [
          "requested_tier",
          "reason"
        ],
        "properties": {
          "requested_tier": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ],
            "description": "Desired plan tier"
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "description": "Justification for the upgrade (10–1000 characters)"
          },
          "requested_overrides": {
            "type": "object",
            "description": "Optional specific limit overrides to request alongside the tier upgrade",
            "properties": {
              "monthly_call_limit": {
                "type": "integer",
                "minimum": 1,
                "description": "Requested monthly API call cap"
              },
              "rate_limit_max_requests": {
                "type": "integer",
                "minimum": 1,
                "description": "Requested per-window rate-limit ceiling"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "PAYMENT_REQUIRED",
          "TOO_MANY_REQUESTS",
          "CONFLICT",
          "INTERNAL_SERVER_ERROR",
          "BAD_GATEWAY",
          "SERVICE_UNAVAILABLE",
          "GATEWAY_TIMEOUT",
          "VALIDATION_ERROR",
          "INVALID_BODY",
          "INVALID_QUERY",
          "INVALID_PARAMS",
          "INVALID_VALUE",
          "GATEWAY_AUTH_CONTEXT_MISSING",
          "UPSTREAM_TARGET_BLOCKED",
          "INSUFFICIENT_BALANCE",
          "SOROBAN_RPC_TIMEOUT",
          "SOROBAN_RPC_ERROR",
          "BILLING_DEDUCTION_FAILED",
          "BILLING_REQUEST_NOT_FOUND",
          "DEVELOPER_NOT_FOUND",
          "API_ACCESS_FORBIDDEN",
          "API_KEY_NOT_FOUND",
          "API_KEY_FORBIDDEN",
          "MISSING_REFRESH_TOKEN",
          "INVALID_REFRESH_TOKEN",
          "REVOKED_TOKEN",
          "EXPIRED_TOKEN",
          "REFRESH_FAILED",
          "REVOKE_FAILED",
          "NOT_AUTHENTICATED",
          "TOKEN_INFO_FAILED",
          "VAULT_NOT_FOUND",
          "VAULT_BALANCE_RETRIEVAL_FAILED",
          "MISSING_AMOUNT",
          "INVALID_AMOUNT_TYPE",
          "INVALID_AMOUNT_FORMAT",
          "INVALID_NETWORK",
          "NETWORK_MISMATCH",
          "INVALID_SOURCE_ACCOUNT",
          "INVALID_TRANSACTION_INPUT",
          "SOURCE_ACCOUNT_NOT_FOUND",
          "INVALID_CONTRACT_ID",
          "NETWORK_UNAVAILABLE",
          "TRANSACTION_BUILD_FAILED",
          "INTERNAL_ERROR",
          "INVALID_WEBHOOK_REGISTRATION",
          "INVALID_WEBHOOK_EVENT_TYPES",
          "WEBHOOK_NOT_FOUND",
          "INVALID_WEBHOOK_URL",
          "WEBHOOK_URL_VALIDATION_FAILED",
          "MISSING_WEBHOOK_SIGNATURE_HEADERS",
          "INVALID_WEBHOOK_TIMESTAMP",
          "WEBHOOK_TIMESTAMP_OUT_OF_WINDOW",
          "MALFORMED_WEBHOOK_SIGNATURE",
          "INVALID_WEBHOOK_SIGNATURE",
          "INVALID_DELIVERY_ID",
          "INVALID_RETRY_POLICY",
          "DLQ_ENTRY_NOT_FOUND",
          "INVALID_IP_FORMAT",
          "IP_NOT_ALLOWED",
          "DATABASE_NOT_AVAILABLE",
          "IDEMPOTENCY_CONFLICT",
          "IDEMPOTENCY_IN_PROGRESS",
          "SIMULATION_FAILED",
          "INVALID_AUTH_HEADER",
          "MISSING_TOKEN",
          "INVALID_TOKEN",
          "MISSING_CLAIMS",
          "TOKEN_EXPIRED",
          "TOKEN_NOT_ACTIVE",
          "QUOTA_REQUEST_NOT_FOUND",
          "QUOTA_REQUEST_ALREADY_RESOLVED",
          "INVALID_QUOTA_REQUEST",
          "REQUEST_TIMEOUT",
          "REQUEST_BODY_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "UNPROCESSABLE_ENTITY",
          "USAGE_AGGREGATE_NOT_FOUND",
          "INVALID_EXPORT_SCHEDULE",
          "EXPORT_SCHEDULE_NOT_FOUND",
          "MISSING_AUTH_FIELDS",
          "AUTH_NOT_IMPLEMENTED",
          "COMPONENT_NOT_CONFIGURED"
        ],
        "description": "Canonical Callora backend error code."
      },
...
    }
  }
}